whisper-dictation/requirements.txt
rumps>=0.4.0
pynput>=1.7.6
sounddevice>=0.4.6
soundfile>=0.12.1
numpy>=1.24.0
openai>=1.0.0
pyperclip>=1.8.2
python-dotenv>=1.0.0
faster-whisper>=1.0.0
pyobjc-core>=10.0
pyobjc-framework-Cocoa>=10.0
pyobjc-framework-Quartz>=10.0
